requirements.txt
pillow
pytesseract
tk
pdf2image
pypdf2
